requirements.txt
aioredis==1.2.0
APScheduler==3.5.3
asgiref==2.3.2
astroid==2.0.4
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
autopep8==1.4.2
certifi==2018.10.15
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
daphne==2.2.2
Django==2.1.2
django-apscheduler==0.2.13
django-markdownx==2.0.25
django-redis==4.9.0
djangorestframework==3.9.0
djangorestframework-composed-permissions==0.2
djangorestframework-jwt==1.11.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
more-itertools==4.3.0
msgpack==0.5.6
Pillow==5.3.0
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.2
pytest==3.9.3
pytz==2018.7
redis==2.10.6
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.12
Twisted==18.9.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.24.1
wrapt==1.10.11
zope.interface==4.6.0
